refactor : Refactor racing winners  - separate racing winners class  - modify class name CarRacingManger to CarRacingGame
