+++ src/main/java/car/racing/CarRacingFactory.java
@@ -11,13 +11,14 @@ public class CarRacingFactory {
 
     private static final String DELIMITER = ",";
 
-    static CarRacingViewController newInstance(String names) {
+    public static CarRacingViewController newInstance(String names) {
         List<Car> cars = Arrays.stream(names.split(DELIMITER))
                 .map(name -> new Car(name, 0))
                 .collect(Collectors.toList());
 
         return new CarRacingViewController(
                 new ResultView(),
-                new CarRacingManager(cars, new CarForwardable()));
+                new CarRacingGame(cars, new CarForwardable()),
+                new CarRacingWinners(cars));
     }
 }
